:recycle: Rename sendV2 to send
