+++ src/components/send/send.test.js
@@ -26,7 +26,7 @@ describe('Form', () => {
     fields: {},
     history: {
       location: {
-        path: '/wallet/sendV2/send',
+        path: '/wallet/send/send',
         search: '?recipient=16313739661670634666L&amount=10&reference=test',
       },
       push: jest.fn(),
@@ -116,7 +116,7 @@ describe('Form', () => {
     },
     history: {
       location: {
-        path: '/wallet/sendV2/send',
+        path: '/wallet/send/send',
         search: '?recipient=16313739661670634666L&amount=10&reference=test',
       },
       push: jest.fn(),
