+++ src/components/send/summary/summary.css
