+++ src/constants/routes.js
@@ -11,7 +11,7 @@ import SingleTransaction from '../components/singleTransaction';
 import HwWalletLogin from '../components/hwWalletLogin';
 // import NotFound from '../components/notFound';
 import AccountVisualDemo from '../components/accountVisual/demo';
-import SendV2 from '../components/sendV2/send';
+import Send from '../components/send/send';
 import Splashscreen from '../components/splashscreen';
 import Register from '../components/register';
 import Login from '../components/login';
@@ -52,7 +52,7 @@ export default {
   },
   send: {
     path: '/wallet/send',
-    component: SendV2,
+    component: Send,
     isPrivate: true,
   },
   wallet: {
